Fuzzy C-means

Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function.

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real,
             [...]) -> FuzzyCMeansResult

Perform Fuzzy C-means clustering over the given data.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

Optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

source

Clustering.FuzzyCMeansResult — Type.

FuzzyCMeansResult{T<:AbstractFloat}

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

source

Missing docstring.

Missing docstring for wcounts(::FuzzyCMeansResult). Check Documenter's build log for details.

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

1000×3 Array{Float64,2}:
 0.334497  0.33401   0.331493
 0.331813  0.337684  0.330503
 0.331747  0.333906  0.334347
 0.333689  0.334181  0.33213 
 0.332384  0.33515   0.332466
 0.33243   0.333606  0.333964
 0.331706  0.335773  0.332522
 0.332995  0.334379  0.332626
 0.333513  0.334275  0.332212
 0.335175  0.331833  0.332992
 ⋮                           
 0.334288  0.333698  0.332014
 0.332024  0.335817  0.332159
 0.332453  0.336499  0.331048
 0.330226  0.335953  0.33382 
 0.331372  0.338455  0.330174
 0.335412  0.332579  0.332009
 0.334452  0.330964  0.334583
 0.331649  0.336055  0.332296
 0.333945  0.331006  0.335049