Fuzzy C-means

Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function.

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real,
             [...]) -> FuzzyCMeansResult

Perform Fuzzy C-means clustering over the given data.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

Optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

source

Clustering.FuzzyCMeansResult — Type.

FuzzyCMeansResult{T<:AbstractFloat} <: ClusteringResult

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

source

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

1000×3 Array{Float64,2}:
 0.332516  0.335849  0.331635
 0.332325  0.336476  0.331199
 0.333651  0.335621  0.330728
 0.333866  0.329477  0.336657
 0.333981  0.329824  0.336195
 0.33482   0.328339  0.336841
 0.332691  0.336894  0.330414
 0.334009  0.327142  0.338849
 0.332759  0.332114  0.335128
 0.331785  0.334679  0.333535
 ⋮                           
 0.33302   0.333956  0.333024
 0.331404  0.338764  0.329831
 0.333874  0.331421  0.334705
 0.333718  0.331502  0.33478 
 0.332323  0.335554  0.332124
 0.333726  0.332353  0.333921
 0.333888  0.331275  0.334837
 0.333359  0.329522  0.337119
 0.33362   0.33326   0.33312