Fuzzy C-means

Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function.

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real; [...])

Performs Fuzzy C-means clustering over the given data.

Returns an instance of FuzzyCMeansResult.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

One may also control the algorithm via the following optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

source

Clustering.FuzzyCMeansResult — Type.

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

source

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

  Iters      center-change
----------------------------
      1       1.119990e+00
      2       4.552338e-03
      3       2.334419e-03
      4       1.842390e-03
      5       1.459606e-03
      6       1.160063e-03
      7       9.247001e-04
Fuzzy C-means converged with 7 iterations (δ = 0.0009247001425853295)
1000×3 Array{Float64,2}:
 0.333412  0.332141  0.334447
 0.332477  0.333392  0.334131
 0.334256  0.333868  0.331876
 0.332387  0.33133   0.336283
 0.332958  0.334283  0.332759
 0.332914  0.33451   0.332576
 0.331875  0.335117  0.333008
 0.332954  0.335149  0.331897
 0.335622  0.333084  0.331294
 0.33727   0.331213  0.331517
 ⋮
 0.332724  0.335588  0.331688
 0.327781  0.334025  0.338194
 0.333996  0.334808  0.331196
 0.332198  0.334901  0.332901
 0.332804  0.332919  0.334277
 0.33209   0.333711  0.3342
 0.333529  0.334308  0.332162
 0.332125  0.332405  0.33547
 0.332805  0.331897  0.335298