Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_\mathcal{C} \ \sum_{i=1}^n \sum_{j=1}^C w_{ij}^\mu \| \mathbf{x}_i - \mathbf{c}_j \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{C} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{\mu-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $\mu > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function

fuzzy_cmeans(data::AbstractMatrix, C::Integer, fuzziness::Real;
             [dist_metric::SemiMetric], [...]) -> FuzzyCMeansResult

Perform Fuzzy C-means clustering over the given data.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Integer: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness ($μ$ in the mathematical formulation), $μ > 1$.

Optional keyword arguments:

dist_metric::SemiMetric (defaults to Euclidean): the SemiMetric object that defines the distance between the data points

maxiter, tol, display, rng: see common options

Clustering.FuzzyCMeansResult — Type

FuzzyCMeansResult{T<:AbstractFloat}

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

Clustering.wcounts — Function

wcounts(R::ClusteringResult) -> Vector{Float64}
wcounts(R::FuzzyCMeansResult) -> Vector{Float64}

Get the weighted cluster sizes as the sum of weights of points assigned to each cluster.

For non-weighted clusterings assumes the weight of every data point is 1.0, so the result is equivalent to convert(Vector{Float64}, counts(R)).

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

1000×3 Matrix{Float64}:
 0.334198  0.331153  0.334649
 0.333496  0.337029  0.329475
 0.331249  0.331492  0.337259
 0.335663  0.331207  0.33313
 0.335574  0.329647  0.334778
 0.331403  0.334458  0.33414
 0.333271  0.327877  0.338852
 0.334267  0.335545  0.330188
 0.331602  0.335148  0.33325
 0.333286  0.329091  0.337623
 ⋮                   
 0.332958  0.330262  0.33678
 0.333958  0.33435   0.331692
 0.333062  0.331471  0.335467
 0.32934   0.336852  0.333808
 0.333705  0.328518  0.337777
 0.33066   0.334105  0.335235
 0.335599  0.330804  0.333597
 0.33375   0.335474  0.330776
 0.330243  0.336941  0.332816