Fuzzy C-means

Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function.

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real; [...])

Performs Fuzzy C-means clustering over the given data.

Returns an instance of FuzzyCMeansResult.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

One may also control the algorithm via the following optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

source

Clustering.FuzzyCMeansResult — Type.

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

source

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

  Iters      center-change
----------------------------
      1       1.114458e+00
      2       6.241864e-03
      3       3.103153e-03
      4       2.360326e-03
      5       1.804574e-03
      6       1.386683e-03
      7       1.100429e-03
      8       8.876436e-04
Fuzzy C-means converged with 8 iterations (δ = 0.0008876435603374225)
1000×3 Array{Float64,2}:
 0.331428  0.33449   0.334082
 0.332405  0.333737  0.333858
 0.334008  0.33389   0.332102
 0.331157  0.334418  0.334425
 0.334117  0.335298  0.330585
 0.334029  0.333613  0.332358
 0.331468  0.335606  0.332927
 0.332695  0.333226  0.334079
 0.334095  0.332073  0.333833
 0.332443  0.333459  0.334099
 ⋮
 0.334189  0.330479  0.335332
 0.333413  0.331994  0.334594
 0.333784  0.334754  0.331463
 0.334325  0.331152  0.334523
 0.333781  0.334837  0.331382
 0.336153  0.331367  0.33248
 0.335954  0.332532  0.331514
 0.334733  0.332662  0.332605
 0.332342  0.333511  0.334147