Fuzzy C-means

Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function.

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real; [...])

Performs Fuzzy C-means clustering over the given data.

Returns an instance of FuzzyCMeansResult.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

One may also control the algorithm via the following optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

source

Clustering.FuzzyCMeansResult — Type.

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

source

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

  Iters      center-change
----------------------------
      1       1.133957e+00
      2       5.709612e-03
      3       2.681123e-03
      4       2.276867e-03
      5       1.942489e-03
      6       1.662129e-03
      7       1.425699e-03
      8       1.225360e-03
      9       1.054930e-03
     10       9.094710e-04
Fuzzy C-means converged with 10 iterations (δ = 0.0009094710281917694)
1000×3 Array{Float64,2}:
 0.331187  0.334113  0.3347
 0.33447   0.331402  0.334129
 0.335301  0.332117  0.332582
 0.332698  0.333532  0.33377
 0.327008  0.337006  0.335987
 0.333102  0.332973  0.333925
 0.327322  0.3357    0.336978
 0.336905  0.331837  0.331258
 0.334545  0.331694  0.333761
 0.331656  0.332989  0.335355
 ⋮
 0.335449  0.331093  0.333457
 0.336363  0.331856  0.331781
 0.333274  0.33417   0.332557
 0.338133  0.330294  0.331573
 0.336786  0.330929  0.332284
 0.335253  0.333114  0.331634
 0.331029  0.332473  0.336498
 0.336135  0.331324  0.332541
 0.33722   0.332207  0.330573