Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real,
             [...]) -> FuzzyCMeansResult

Perform Fuzzy C-means clustering over the given data.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

Optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

Clustering.FuzzyCMeansResult — Type

FuzzyCMeansResult{T<:AbstractFloat}

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

Missing docstring.

Missing docstring for wcounts(::FuzzyCMeansResult). Check Documenter's build log for details.

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

1000×3 Matrix{Float64}:
 0.337029  0.333744  0.329227
 0.331497  0.331928  0.336575
 0.334665  0.331756  0.333579
 0.3352    0.333102  0.331698
 0.330351  0.335524  0.334124
 0.334844  0.332561  0.332594
 0.331722  0.332295  0.335982
 0.336116  0.334997  0.328887
 0.334237  0.335953  0.32981
 0.337261  0.333579  0.329161
 ⋮                   
 0.330954  0.332957  0.336088
 0.333939  0.335971  0.33009
 0.334949  0.333015  0.332036
 0.332767  0.332871  0.334362
 0.334493  0.33283   0.332677
 0.33455   0.331163  0.334287
 0.330844  0.332555  0.336601
 0.335427  0.332414  0.332159
 0.332244  0.330727  0.337029