Fuzzy C-means

Fuzzy C-means is a clustering method that provides cluster membership weights instead of "hard" classification (e.g. K-means).

From a mathematical standpoint, fuzzy C-means solves the following optimization problem:

\[\arg\min_C \ \sum_{i=1}^n \sum_{j=1}^c w_{ij}^m \| \mathbf{x}_i - \mathbf{c}_{j} \|^2, \ \text{where}\ w_{ij} = \left(\sum_{k=1}^{c} \left(\frac{\left\|\mathbf{x}_i - \mathbf{c}_j \right\|}{\left\|\mathbf{x}_i - \mathbf{c}_k \right\|}\right)^{\frac{2}{m-1}}\right)^{-1}\]

Here, $\mathbf{c}_j$ is the center of the $j$-th cluster, $w_{ij}$ is the membership weight of the $i$-th point in the $j$-th cluster, and $m > 1$ is a user-defined fuzziness parameter.

Clustering.fuzzy_cmeans — Function

fuzzy_cmeans(data::AbstractMatrix, C::Int, fuzziness::Real,
             [...]) -> FuzzyCMeansResult

Perform Fuzzy C-means clustering over the given data.

Arguments

data::AbstractMatrix: $d×n$ data matrix. Each column represents one $d$-dimensional data point.

C::Int: the number of fuzzy clusters, $2 ≤ C < n$.

fuzziness::Real: clusters fuzziness (see $m$ in the mathematical formulation), $\mathrm{fuzziness} > 1$.

Optional keyword arguments:

dist_metric::Metric (defaults to Euclidean): the Metric object that defines the distance between the data points

maxiter, tol, display: see common options

Clustering.FuzzyCMeansResult — Type

FuzzyCMeansResult{T<:AbstractFloat}

The output of fuzzy_cmeans function.

Fields

centers::Matrix{T}: the $d×C$ matrix with columns being the centers of resulting fuzzy clusters

weights::Matrix{Float64}: the $n×C$ matrix of assignment weights ($\mathrm{weights}_{ij}$ is the weight (probability) of assigning $i$-th point to the $j$-th cluster)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

Missing docstring.

Missing docstring for wcounts(::FuzzyCMeansResult). Check Documenter's build log for details.

Examples

using Clustering

# make a random dataset with 1000 points
# each point is a 5-dimensional vector
X = rand(5, 1000)

# performs Fuzzy C-means over X, trying to group them into 3 clusters
# with a fuzziness factor of 2. Set maximum number of iterations to 200
# set display to :iter, so it shows progressive info at each iteration
R = fuzzy_cmeans(X, 3, 2, maxiter=200, display=:iter)

# get the centers (i.e. weighted mean vectors)
# M is a 5x3 matrix
# M[:, k] is the center of the k-th cluster
M = R.centers

# get the point memberships over all the clusters
# memberships is a 20x3 matrix
memberships = R.weights

1000×3 Matrix{Float64}:
 0.331555  0.331727  0.336718
 0.333702  0.335582  0.330716
 0.332718  0.334794  0.332488
 0.331795  0.333041  0.335165
 0.330384  0.334284  0.335333
 0.337766  0.330791  0.331443
 0.337155  0.329695  0.33315
 0.335721  0.329748  0.334531
 0.335291  0.328609  0.3361
 0.335135  0.331603  0.333262
 ⋮                   
 0.335824  0.33138   0.332796
 0.330914  0.333983  0.335103
 0.335222  0.33463   0.330148
 0.336529  0.333155  0.330316
 0.335739  0.332692  0.331569
 0.332233  0.33501   0.332757
 0.330541  0.334227  0.335233
 0.336139  0.330336  0.333525
 0.331076  0.336354  0.33257