K-means

K-means

K-means is a classical method for clustering or vector quantization. It produces a fixed number of clusters, each associated with a center (also known as a prototype), and each data point is assigned to a cluster with the nearest center.

From a mathematical standpoint, K-means is a coordinate descent algorithm that solves the following optimization problem:

\[\text{minimize} \ \sum_{i=1}^n \| \mathbf{x}_i - \boldsymbol{\mu}_{z_i} \|^2 \ \text{w.r.t.} \ (\boldsymbol{\mu}, z)\]

Here, $\boldsymbol{\mu}_k$ is the center of the $k$-th cluster, and $z_i$ is an index of the cluster for $i$-th point $\mathbf{x}_i$.

Clustering.kmeans — Function.

kmeans(X, k, [...])

K-means clustering of the $d \times n$ data matrix X (each column of X is a $d$-dimensional data point) into $k$ clusters.

Returns KmeansResult object.

Algorithm Options

init (defaults to :kmpp): how cluster seeds should be initialized, could be one of the following:

a Symbol, the name of a seeding algorithm (see Seeding for a list of supported methods).

an integer vector of length $k$ that provides the indices of points to use as initial seeds.

weights: $n$-element vector of point weights (the cluster centers are the weighted means of cluster members)

maxiter, tol, display: see common options

source

Clustering.KmeansResult — Type.

The output of K-means algorithm.

See also: kmeans, kmeans!.

source

If you already have a set of initial center vectors, kmeans! could be used:

Clustering.kmeans! — Function.

kmeans!(X, centers; [kwargs...])

Update the current cluster centers ($d \times k$ matrix, where $d$ is the dimension and $k$ the number of centroids) using the $d \times n$ data matrix X (each column of X is a $d$-dimensional data point).

Returns KmeansResult object.

See kmeans for the description of optional kwargs.

source

Examples

using Clustering

# make a random dataset with 1000 random 5-dimensional points
X = rand(5, 1000)

# cluster X into 20 clusters using K-means
R = kmeans(X, 20; maxiter=200, display=:iter)

@assert nclusters(R) == 20 # verify the number of clusters

a = assignments(R) # get the assignments of points to clusters
c = counts(R) # get the cluster sizes
M = R.centers # get the cluster centers

  Iters               objv        objv-change | affected
-------------------------------------------------------------
      0       2.092736e+02
      1       1.490465e+02      -6.022706e+01 |       20
      2       1.378416e+02      -1.120494e+01 |       20
      3       1.333974e+02      -4.444185e+00 |       20
      4       1.313504e+02      -2.047069e+00 |       20
      5       1.298069e+02      -1.543440e+00 |       19
      6       1.287397e+02      -1.067191e+00 |       15
      7       1.280882e+02      -6.515025e-01 |       18
      8       1.277250e+02      -3.632440e-01 |       15
      9       1.271662e+02      -5.587861e-01 |       14
     10       1.265064e+02      -6.597583e-01 |       13
     11       1.258256e+02      -6.808056e-01 |       14
     12       1.255205e+02      -3.051365e-01 |       11
     13       1.254046e+02      -1.159215e-01 |        6
     14       1.253549e+02      -4.969517e-02 |        2
     15       1.253388e+02      -1.605279e-02 |        3
     16       1.252972e+02      -4.162405e-02 |        4
     17       1.252859e+02      -1.133964e-02 |        0
     18       1.252859e+02       0.000000e+00 |        0
K-means converged with 18 iterations (objv = 125.28585762840214)
5×20 Array{Float64,2}:
 0.722636  0.25834   0.258741  0.732108  …  0.772569  0.240483  0.283344
 0.74265   0.735661  0.183422  0.203601     0.702998  0.435551  0.642808
 0.836331  0.315752  0.361552  0.707368     0.273039  0.706107  0.177199
 0.793645  0.302533  0.707575  0.396807     0.185319  0.155915  0.841191
 0.374155  0.20338   0.206613  0.80534      0.23588   0.308636  0.493772

using RDatasets, Clustering, Plots
iris = dataset("datasets", "iris"); # load the data

features = collect(Matrix(iris[:, 1:4])'); # features to use for clustering
result = kmeans(features, 3); # run K-means for the 3 clusters

# plot with the point color mapped to the assigned cluster index
scatter(iris.PetalLength, iris.PetalWidth, marker_z=result.assignments,
        color=:lightrainbow, legend=false)