K-medoids

K-medoids

K-medoids is a clustering algorithm that works by finding $k$ data points (called medoids) such that the total cost or total distance between each data point and the closest medoid is minimal.

Clustering.kmedoids — Function.

kmedoids(costs::DenseMatrix, k::Integer; ...) -> KmedoidsResult

Perform K-medoids clustering of $n$ points into k clusters, given the costs matrix ($n×n$, costs[i, j] is the cost of assigning j-th point to the medoid represented by the i-th point).

Note

This package implements a K-means style algorithm instead of PAM, which is considered much more efficient and reliable.

Arguments

init (defaults to :kmpp): how medoids should be initialized, could be one of the following:

a Symbol indicating the name of a seeding algorithm (see Seeding for a list of supported methods).

an integer vector of length k that provides the indices of points to use as initial medoids.

maxiter, tol, display: see common options

source

Clustering.kmedoids! — Function.

kmedoids!(costs::DenseMatrix, medoids::Vector{Int};
          [kwargs...]) -> KmedoidsResult

Update the current cluster medoids using the costs matrix.

The medoids field of the returned KmedoidsResult points to the same array as medoids argument.

See kmedoids for the description of optional kwargs.

source

Clustering.KmedoidsResult — Type.

KmedoidsResult{T} <: ClusteringResult

The output of kmedoids function.

Fields

medoids::Vector{Int}: the indices of $k$ medoids

assignments::Vector{Int}: the indices of clusters the points are assigned to, so that medoids[assignments[i]] is the index of the medoid for the $i$-th point

acosts::Vector{T}: assignment costs, i.e. acosts[i] is the cost of assigning $i$-th point to its medoid

counts::Vector{Int}: cluster sizes

totalcost::Float64: total assignment cost (the sum of acosts)

iterations::Int: the number of executed algorithm iterations

converged::Bool: whether the procedure converged

source